Move BundleGraph to core
